make DumpIndent take a uint8_t

+++ universe/UniverseObject.h
@@ -274,7 +274,7 @@ private:
 
 /** A function that returns the correct amount of spacing for an indentation of
   * \p ntabs during a dump. */
-[[nodiscard]] inline std::string DumpIndent(unsigned short ntabs = 1)
+[[nodiscard]] inline std::string DumpIndent(uint8_t ntabs = 1)
 { return std::string(ntabs * 4 /* conversion to std::size_t is safe */, ' '); }
